Add header to searchLocationReverse (#458)
As described in #458, uploading trails intermittently fails because the request to the Nominatim API in `searchLocationReverse` does not
contain a header as required by the API. The header was added to `searchLocations` in 47f431e, but the issue with the trail upload was
caused by it missing in `searchLocationReverse`. This PR adds it there as well.